requirements.txt
altgraph==0.17.4
ansys-api-mapdl==0.5.2
ansys-api-platform-instancemanagement==1.1.0
ansys-dpf-core==0.12.2
ansys-dpf-post==0.8.0
ansys-mapdl-core==0.68.4
ansys-mapdl-reader==0.53.0
ansys-math-core==0.1.5
ansys-platform-instancemanagement==1.1.2
ansys-tools-path==0.6.0
appdirs==1.4.4
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
geomdl==5.3.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
httplib2==0.22.0
idna==3.7
importlib_metadata==8.0.0
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==1.26.4
numpy-stl==3.1.1
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyansys-tools-versioning==0.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyiges==0.3.1
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-utils==3.8.2
pyvista==0.44.0
pywin32-ctypes==0.2.2
requests==2.32.3
rsa==4.9
scipy==1.14.0
scooby==0.10.0
six==1.16.0
tabulate==0.9.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vtk==9.3.1
zipp==3.19.2